+++ src/include/win/WindowsDRAMPowerReader.h
@@ -52,11 +52,11 @@ namespace xmem {
 		/**
 		 * @brief Constructor.
 		 * @param counter_cpu_index Which CPU's DRAM power counter to sample. A single hardware counter might be shared across different CPUs.
-		 * @param sampling_period The time between power samples in seconds.
+		 * @param sampling_period The time between power samples in milliseconds.
 		 * @param power_units The power units for each sample in watts.
 		 * @param cpu_affinity The CPU affinity for this object's run() method for any thread that calls it. If negative, no affinity preference.
 		 */
-		WindowsDRAMPowerReader(uint32_t counter_cpu_index, double sampling_period, double power_units, std::string name, int32_t cpu_affinity);
+		WindowsDRAMPowerReader(uint32_t counter_cpu_index, uint64_t sampling_period, double power_units, std::string name, int32_t cpu_affinity);
 
 		/**
 		 * @brief Destructor.
